IQSS/6829 - create mode uses DatasetPage - initialize identifier there
for directUpload case
